# backend/openapi-baseline.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RateSession": {
        "properties": {
          "satisfaction": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Satisfaction",
            "type": "integer"
          }
        },
        "required": [
          "satisfaction"
        ],
        "title": "RateSession",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "跨境电商赛训效率工具箱",
    "version": "1.8.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ai-chat/session/{session_id}/rate": {
      "post": {
        "description": "满意度评分",
        "operationId": "rate_session_api_ai_chat_session__session_id__rate_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rate Session",
        "tags": [
          "AI客服"
        ]
      }
    },
...
    "/api/freight-rate-packs": {
      "get": {
        "operationId": "list_rate_releases_api_freight_rate_packs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Rate Releases",
        "tags": [
          "物流费率版本"
        ]
      }
    },
    "/api/freight-rate-packs/current": {
      "get": {
        "operationId": "get_current_rate_pack_api_freight_rate_packs_current_get",
        "parameters": [
          {
            "in": "query",
            "name": "pack_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Current Rate Pack",
        "tags": [
          "物流费率版本"
        ]
      }
    },
    "/api/freight-rate-packs/drafts": {
      "post": {
        "operationId": "create_draft_api_freight_rate_packs_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Draft",
        "tags": [
          "物流费率版本"
        ]
      }
    },
    "/api/freight-rate-packs/{pack_id}/rollback": {
      "post": {
        "operationId": "rollback_rate_pack_api_freight_rate_packs__pack_id__rollback_post",
        "parameters": [
          {
            "in": "path",
            "name": "pack_id",
            "required": true,
            "schema": {
              "title": "Pack Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rollback Rate Pack",
        "tags": [
          "物流费率版本"
        ]
      }
    },
    "/api/freight-rate-packs/{pack_id}/{version}/artifact": {
      "get": {
        "operationId": "get_rate_artifact_api_freight_rate_packs__pack_id___version__artifact_get",
        "parameters": [
          {
            "in": "path",
            "name": "pack_id",
            "required": true,
            "schema": {
              "title": "Pack Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Rate Artifact",
        "tags": [
          "物流费率版本"
        ]
      }
    },
    "/api/freight-rate-packs/{pack_id}/{version}/publish": {
      "post": {
        "operationId": "publish_rate_pack_api_freight_rate_packs__pack_id___version__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "pack_id",
            "required": true,
            "schema": {
              "title": "Pack Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Rate Pack",
        "tags": [
          "物流费率版本"
        ]
      }
    },
...
    "/api/help/faq/list": {
      "get": {
        "description": "获取 FAQ 列表\n\n支持按平台、功能、分类筛选",
        "operationId": "get_faq_list_api_help_faq_list_get",
        "parameters": [
          {
            "in": "query",
            "name": "platform_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform Key"
            }
          },
          {
            "in": "query",
            "name": "capability_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Capability Key"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Faq List",
        "tags": [
          "帮助查询"
        ]
      }
    },
    "/api/help/query": {
      "post": {
        "description": "智能帮助查询 - FAQ 优先，AI 兜底\n\n请求参数:\n- platform_key: 平台标识 (amazon/aliexpress)\n- capability_key: 功能标识 (register/logistics_template/listing/ads/shipping/fba_agl)\n- question: 用户问题\n\n返回:\n- matched: 是否匹配到 FAQ\n- ai_used: 是否使用了 AI\n- source: 答案来源 (faq/ai)\n- answer: 答案内容\n- faq_id: FAQ ID (如果匹配)\n- need_ai: 是否需要 AI 兜底",
        "operationId": "query_help_api_help_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Request",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Query Help",
        "tags": [
          "帮助查询"
        ]
      }
    },
...
  }
}